Move config to app/config package and adopt Cloud Native env var approach
- Replace hardcoded values in config files with os.getenv() calls
- Move app/data/config.py → app/config/__init__.py
- Move app/data/currates_conf.py → app/config/currates_conf.py
- Delete examples/config.py and examples/currates_conf.py (superseded by .env.example)
- Add .env.example documenting all environment variables
- Remove app/data volume mount from docker-compose.yml; mount only currates_conf.py for runtime editability without rebuild
- Update init.sh and start-container.sh to use .env instead of config file copies
- Load .env in start.sh for non-Docker usage
