+++ .env.example
@@ -1,0 +1,26 @@
+# Telegram bot token (required)
+TOKEN=
+
+# Currency rates data sources API keys (required)
+EXCHANGE_RATE_API_KEY=
+COINMARKETCAP_API_KEY=
+
+# Bot identity & webhook (HOST required only when DEBUG=false)
+HOST=
+BOT_NAME=textUtilsBot
+SERVER_PORT=8443
+
+# Local server
+APP_HOST=0.0.0.0
+APP_PORT=8080
+METRICS_PORT=8000
+
+# Unix socket (alternative to TCP)
+UNIX_SOCKET=/tmp/textUtilsBot.sock
+SOCKET_TYPE=TCP
+
+# Optional proxy for Telegram API access
+PROXY=
+
+# Polling mode for local dev; false = webhooks in production
+DEBUG=true
