+++ start.sh
@@ -1,4 +1,5 @@
 #!/bin/sh
 
+[ -f .env ] && . .env
 . venv/bin/activate
 exec python -m app.bot
